+++ backend/src/main/java/com/aqryuz/backend/authentication/util/JwtUtils.java
@@ -1,24 +1,28 @@
 package com.aqryuz.backend.authentication.util;
 
-import com.aqryuz.backend.authentication.config.JWTProperties;
-import com.aqryuz.backend.authentication.model.User;
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.JwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.security.Keys;
 import java.util.Date;
+
 import javax.crypto.SecretKey;
-import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
+import com.aqryuz.backend.authentication.config.JWTProperties;
+import com.aqryuz.backend.authentication.model.User;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.JwtException;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.security.Keys;
+import lombok.extern.slf4j.Slf4j;
+
 @Component
 @Slf4j
 public class JwtUtils {
 
-  private final int jwtExpirationMs;
+  private final Long jwtExpirationMs;
 
   private final SecretKey jwtSecret; // Store as SecretKey
